+++ plugins/postcss-color-functional-notation/src/has-supports-at-rule-ancestor.ts
@@ -1,6 +1,6 @@
 import type { Node, AtRule } from 'postcss';
 
-const supportsCheck = /(\(color: rgb(a?)\(0 0 0 \/ 0)|(\(color: hsl(a?)\(0 0% 0% \/ 0)/i;
+export const rgb_hsl_functionRegex = /(?:rgb|hsl)a?\(/i;
 
 export function hasSupportsAtRuleAncestor(node: Node): boolean {
 	let parent = node.parent;
@@ -10,8 +10,10 @@ export function hasSupportsAtRuleAncestor(node: Node): boolean {
 			continue;
 		}
 
-		if ((parent as AtRule).name === 'supports' && supportsCheck.test((parent as AtRule).params)) {
-			return true;
+		if ((parent as AtRule).name.toLowerCase() === 'supports') {
+			if (rgb_hsl_functionRegex.test((parent as AtRule).params)) {
+				return true;
+			}
 		}
 
 		parent = parent.parent;